[AT] AdaptiveLearning implementation: Concept elements CRUD: Update

+++ modules/complex/src/app/complexMaterial/components/materials/add-material-popover/base-file-management.component.ts
@@ -51,8 +51,10 @@ export class BaseFileManagementComponent<T> implements OnInit, OnDestroy {
     if (toSave) {
       this.onSave(files);
     } else {
-      files.filter(f => f.IdFile <= 0)
-        .forEach(f => this.deleteFile(f));
+      if (files) {
+        files.filter(f => f.IdFile <= 0)
+          .forEach(f => this.deleteFile(f));
+      }
       this.dialogRef.close();
     }
   }
